requirements.txt
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
defusedxml==0.7.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
fpdf==1.7.2
fpdf2==2.8.3
frontend==0.0.3
gensim==4.3.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
jellyfish==1.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
looseversion==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
msgspec==0.19.0
murmurhash==1.0.13
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==1.26.4
openai==1.88.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
preshed==3.0.10
proto-plus==1.26.1
protobuf==5.29.5
prov==2.0.1
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
Pygments==2.19.1
PyMuPDF==1.26.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
rsa==4.9.1
Sastrawi==1.0.1
scikit-learn==1.6.1
scipy==1.13.1
segtok==1.5.11
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.1
starlette==0.46.2
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.6.0
tools==1.0.2
tqdm==4.67.1
traits==7.0.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
utils==1.0.2
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wikipedia==1.4.0
wordcloud==1.9.4
wrapt==1.17.2
yake==0.4.8
